langchain_processor/requirements.txt
langchain
python-dotenv
langchain_groq
langchain_community
langchain-core
pandas
pydantic
beautifulsoup4
tqdm
langchain-huggingface